Move orderSummary into it's own file from checkOut.js to make the code organised
